Chapter 5 : Code practise

+++ src/Chapter4_ObjectBehavior/TestMethod.java
@@ -3,14 +3,13 @@ package Chapter4_ObjectBehavior;
 public class TestMethod {
 
 	public static void main(String[] args) {
-		TestMethod t  = new TestMethod();
-		long c =7 ;
+		TestMethod t = new TestMethod();
+		short c = 7; // short 隱式轉換為int
 		int value = t.calcArea(c, 15);
 		System.out.println(value);
-
 	}
 
-	public int calcArea  (int height  , int width) {
+	public int calcArea(int height, int width) {
 		return height * width;
 	}
